+++ client/src/components/Map/ResultList.js
