+++ client/src/components/Map/SimpleMap.js
