+++ client/src/components/games/SimpleGameTile.js
@@ -13,5 +13,4 @@ const SimpleGameTile = ({ game }) => {
   )
 }
 
-
 export default SimpleGameTile
