+++ client/src/components/layout/TopBar.js
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { Link } from "react-router-dom";
 import SignOutButton from "../authentication/SignOutButton";
-import UserTile from "../UserTile";
+import UserTile from "../users/UserTile";
 
 const TopBar = ({ user }) => {
   const unauthenticatedListItems = [
