+++ client/src/components/plans/NewPlanForm.js
@@ -1,38 +1,33 @@
-import React, { useEffect, useState } from 'react';
-import { Redirect, Route } from 'react-router-dom';
-import { format } from 'date-fns'
+import React, { useState } from 'react';
+import { Redirect } from 'react-router-dom';
+import translateServerErrors from '../../services/translateServerErrors'
+import FormError from '../layout/FormError';
+import GameSearchField from '../games/GameSearchField';
+import SelectedGameTile from '../games/SelectedGameTile';
+import GoogleMap from "../Map/GoogleMap"
+import ErrorList from '../layout/ErrorList';
 
-import GameSearchField from './GameSearchField';
-import SelectedGameTile from './SelectedGameTile';
-import FormError from './layout/FormError';
-import PlanShowPage from './PlanShowPage';
-import GoogleMap from './GoogleMap';
-import ErrorList from './layout/ErrorList';
-import translateServerErrors from '../services/translateServerErrors';
 
-const EditPlanForm = ({ user, plan, setPlan }) => {
+const NewPlan = (props) => {
+  const { user } = props
+  const { id } = props.match.params
+
   const [formSuccess, setFormSuccess] = useState(false)
-  const [deleteSuccess, setDeleteSuccess] = useState(false)
+  const [formData, setFormData] = useState({})
   const [errors, setErrors] = useState({})
   const [game, setGame] = useState({})
   const [showMap, setShowMap] = useState(false)
 
-  const [formData, setFormData] = useState({})
-
-  useEffect(() => {
-    const dateObject = new Date(plan.date)
-    const formattedDate = format(dateObject, 'yyyy-MM-dd')
-    const formattedTime = format(dateObject, 'HH:mm')
-    setFormData({
-      name: plan.name,
-      date: formattedDate,
-      time: formattedTime,
-      location: plan.location,
-      address: plan.address,
-      gameId: plan.game.id
-    })
-
-    setGame(plan.game)
+  useState(async () => {
+    if (id) {
+      try {
+        const response = await fetch(`/api/v1/games/${id}`)
+        const body = await response.json()
+        setGame(body.game)
+      } catch (error) {
+        console.error(error)
+      }
+    }
   }, [])
 
   const handleChange = (event) => {
@@ -42,40 +37,25 @@ const EditPlanForm = ({ user, plan, setPlan }) => {
     })
   }
 
-  const handleSubmit = async (event) => {
+  const handleSubmit = async event => {
     event.preventDefault()
     try {
-      const response = await fetch(`/api/v1/plans/${plan.id}`, {
-        method: "PATCH",
+      const response = await fetch("/api/v1/plans", {
+        method: "POST",
         headers: new Headers({
           "Content-Type": "application/json"
         }),
         body: JSON.stringify({ ...formData, gameId: game.id })
       })
       if (!response.ok) {
-        const body = await response.json()
-        const newErrors = translateServerErrors(body.errors)
-        return setErrors(newErrors)
-      } else {
-        const body = await response.json()
-        setFormSuccess(true)
+        if (response.status === 422) {
+          const body = await response.json()
+          const newErrors = translateServerErrors(body.errors)
+          return setErrors(newErrors)
+        }
+        throw new Error(response.statusText)
       }
-    } catch (error) {
-      console.error(`error in fetch: ${error}`)
-    }
-  }
-
-  const handleDelete = async (event) => {
-    event.preventDefault()
-    try {
-      const response = await fetch(`/api/v1/plans/${plan.id}`, {
-        method: "DELETE"
-      })
-      const body = await response.json()
-      if (body.success) {
-        return setDeleteSuccess(true)
-      }
-      throw new Error(body)
+      setFormSuccess(true)
     } catch (error) {
       console.error(error)
     }
@@ -102,10 +82,13 @@ const EditPlanForm = ({ user, plan, setPlan }) => {
 
   const setSelectedGame = (game) => {
     setGame(game)
-    setFormData({ ...formData, gameId: game.id })
     if (!game.id) {
       addSelectedGame(game)
     }
+  }
+
+  if (formSuccess) {
+    return <Redirect to="/plans" />
   }
 
   let gameSelectComponent
@@ -115,16 +98,7 @@ const EditPlanForm = ({ user, plan, setPlan }) => {
     gameSelectComponent = <SelectedGameTile game={game} setGame={setGame} />
   }
 
-  if (deleteSuccess) {
-    return <Redirect to="/plans" />
-  }
-
-  if (formSuccess) {
-    return <Route exact path="/plans/:id"
-      render={props => <PlanShowPage {...props} user={user} />} />
-  }
-
-  if (user.id !== plan.owner.id) {
+  if (!user) {
     return <p>Not authorized</p>
   }
 
@@ -149,6 +123,7 @@ const EditPlanForm = ({ user, plan, setPlan }) => {
             value={formData.location}
             onChange={handleChange}
           />
+          <FormError error={errors.Location} />
         </label>
         <label>Address:
           <input
@@ -157,6 +132,7 @@ const EditPlanForm = ({ user, plan, setPlan }) => {
             value={formData.address}
             onChange={handleChange}
           />
+          <FormError error={errors.Location} />
         </label>
         <button className='button' onClick={event => {
           event.preventDefault()
@@ -167,7 +143,8 @@ const EditPlanForm = ({ user, plan, setPlan }) => {
 
   return (
     <form className='grid-container' onSubmit={handleSubmit}>
-      <h4>Edit plan ({plan.id}):</h4>
+      <h4>Add a new plan</h4>
+      <ErrorList errors={errors} />
       <label>Name:
         <input
           type="text"
@@ -196,18 +173,16 @@ const EditPlanForm = ({ user, plan, setPlan }) => {
           </label>
         </div>
       </label>
-      <label>
+      <FormError error={errors.Date} />
+      <div className='grid-x'>
         {mapComponent}
-        <FormError error={errors.Location} />
-      </label>
+
+      </div>
       {gameSelectComponent}
       {errors["Game Id"] ? <FormError error={errors["Game Id"]} /> : ""}
-      <div className='edit-delete'>
-        <button className='button'>Update</button>
-        <button className='button alert' onClick={handleDelete}>Delete</button>
-      </div>
+      <button className='button'>Submit</button>
     </form>
   )
 }
 
-export default EditPlanForm
+export default NewPlan
